requirements.txt
altgraph==0.17.4
Babel==2.14.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==23.2
pefile==2023.2.7
pyasn1==0.5.1
pygame==2.5.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
PySimpleGUI==5.0.2
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.3
rsa==4.9
setuptools==69.1.1
SQLAlchemy==2.0.32
tk==0.1.0
tkcalendar==1.6.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
gunicorn